requirements.txt
streamlit
gspread
oauth2client
google-api-python-client
google-cloud-storage
fastapi
uvicorn
gunicorn
pandas
scikit-learn
joblib
python-multipart
plotly
wordcloud
matplotlib